+++ src/App.jsx
@@ -1,0 +1,13 @@
+
+import './App.css'
+import CharCounter from './components/CharCounter';
+
+function App() {
+
+    return (
+        <div>
+            <CharCounter />
+        </div>
+    )
+}
+export default App;
