+++ src/main.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import CharCounter from './CharCounter.jsx'
+import App from './App.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <CharCounter />
+    <App />
   </React.StrictMode>,
 )
